rename package, remove old stuffs
