+++ anna/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "anna-client-tokio"
+name = "wasmedge-anna-driver"
 version = "0.1.0"
 edition = "2021"
 license = "Apache-2.0"
@@ -21,14 +21,6 @@ futures-timer = "3.0.2"
 once_cell = "1.8.0"
 serde_json = "1.0.64"
 anna-api = { path = "api" }
-# tokio = { version = "1", features = [
-#     "rt",
-#     "net",
-#     "time",
-#     "sync",
-#     "macros",
-#     "io-util",
-# ] }
 tokio = { git = "https://github.com/WasmEdge/tokio.git", branch = "wasmedge", features = [
     "rt",
     "net",
